+++ packages/gatsby-plugin-orion-edit/src/components/MenuCard/MenuCard.test.js
@@ -3,22 +3,18 @@ import React from 'react'
 import MenuCard from '.'
 import { render } from '@testing-library/react'
 
+const MockImage = () => <div>mock image</div>
+
 const renderComponent = () =>
-  render(<MenuCard src="test.jpg" label="Test Item" to="me" />)
+  render(<MenuCard Image={MockImage} label="Test Item" to="me" />)
 
 describe('MenuCard component', () => {
   it('renders a text label', () => {
     const { getByText } = renderComponent()
     expect(getByText('Test Item')).toBeInTheDocument()
   })
-  it('renders an image using the MUI card media component', () => {
-    const { getByTitle } = renderComponent()
-    expect(getByTitle('Test Item')).toMatchInlineSnapshot(`
-      <div
-        class="MuiCardMedia-root makeStyles-media-85"
-        style="background-image: url(test.jpg);"
-        title="Test Item"
-      />
-    `)
+  it('renders the image component from props', () => {
+    const { getByText } = renderComponent()
+    expect(getByText('mock image')).toBeInTheDocument()
   })
 })