+++ packages/gatsby-plugin-orion-edit/src/components/MenuCard/index.js
@@ -5,7 +5,6 @@ import { makeStyles } from '@material-ui/core'
 import Card from '@material-ui/core/Card'
 import CardActionArea from '@material-ui/core/CardActionArea'
 import CardContent from '@material-ui/core/CardContent'
-import CardMedia from '@material-ui/core/CardMedia'
 import Typography from '@material-ui/core/Typography'
 
 const cardSize = 256
@@ -24,20 +23,39 @@ const useStyles = makeStyles(theme => ({
   media: {
     minHeight: 200,
   },
+  content: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+  },
+  title: {
+    marginTop: '16px',
+  },
+  image: {
+    minHeight: '184px',
+    display: 'flex',
+    alignItems: 'flex-end',
+  },
 }))
-const MenuCard = ({ label, src, to }) => {
+const MenuCard = ({ label, Image, to }) => {
   const classes = useStyles()
   return (
     <Card className={classes.root}>
       <CardActionArea>
         <Link to={to} className={classes.link}>
-          <CardContent>
-            <CardMedia className={classes.media} image={src} title={label} />
+          <CardContent className={classes.content}>
+            {Image && (
+              <div className={classes.image}>
+                <Image />
+              </div>
+            )}
             <Typography
               align="center"
               color="secondary"
               variant="h5"
               component="h5"
+              className={classes.title}
             >
               {label}
             </Typography>
@@ -51,7 +69,7 @@ const MenuCard = ({ label, src, to }) => {
 MenuCard.propTypes = {
   label: PropTypes.string.isRequired,
   to: PropTypes.string.isRequired,
-  src: PropTypes.string.isRequired,
+  Image: PropTypes.node.isRequired,
 }
 
 export default MenuCard