+++ packages/acme-edit/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.46.0](https://github.com/nearform/orion/compare/v2.45.0...v2.46.0) (2020-04-17)
+
+**Note:** Version bump only for package acme-edit
+
+
+
+
+
+# [2.45.0](https://github.com/nearform/orion/compare/v2.44.0...v2.45.0) (2020-04-15)
+
+**Note:** Version bump only for package acme-edit
+
+
+
+
+
+# [2.44.0](https://github.com/nearform/orion/compare/v2.43.0...v2.44.0) (2020-04-10)
+
+**Note:** Version bump only for package acme-edit
+
+
+
+
+
 # [2.43.0](https://github.com/nearform/orion/compare/v2.42.0...v2.43.0) (2020-04-09)
 
 **Note:** Version bump only for package acme-edit